moved proto factory into proto dir; removed old fed env parser and utils in encryption
