+++ metisfl/driver/controller_client.py
@@ -1,7 +1,7 @@
 from metisfl.grpc.grpc_services import GRPCClient
 from metisfl.proto import controller_pb2, service_common_pb2
 from metisfl.utils.metis_logger import MetisLogger
-from metisfl.utils.proto_messages_factory import ModelProtoMessages
+from metisfl.proto.proto_messages_factory import ModelProtoMessages
 
 
 # FIXME(@stripeli) - logic here implies that requests go through without errors
