+++ metisfl/learner/learner.py
@@ -2,7 +2,7 @@
 import metisfl.proto.metis_pb2 as metis_pb2
 from metisfl import config
 from metisfl.models.utils import get_model_ops_fn
-from metisfl.utils.proto_messages_factory import MetisProtoMessages
+from metisfl.proto.proto_messages_factory import MetisProtoMessages
 
 from .dataset_handler import LearnerDataset
 from .learner_executor import LearnerExecutor
