+++ metisfl/proto/model.proto
@@ -114,40 +114,7 @@ message FederatedModel {
 ////////////////
 
 message OptimizerConfig {
-  oneof config {
-    VanillaSGD vanilla_sgd = 1;
-    MomentumSGD momentum_sgd = 2;
-    FedProx fed_prox = 3;
-    Adam adam = 4;
-    AdamWeightDecay adam_weight_decay = 5;
-    Custom custom = 6;
-  }
-}
-
-message VanillaSGD {
-  float learning_rate = 1;
-  float L1_reg = 2;
-  float L2_reg = 3;
-}
-
-message MomentumSGD {
-  float learning_rate = 1;
-  float momentum_factor = 2;
-}
-
-message FedProx {
-  float learning_rate = 1;
-  float proximal_term = 2;
-}
-
-message Adam {
-  float learning_rate = 1;
-  float beta_1 = 2;
-  float beta_2 = 3;
-  float epsilon = 4;
-}
-
-message AdamWeightDecay {
-  float learning_rate = 1;
-  float weight_decay = 2;
+  string name = 1;
+  string learning_rate = 2;
+  map<string, string> kwargs = 3;
 }
