+++ test/learner_notrain_noeval.py
@@ -5,7 +5,7 @@ import time
 import cloudpickle
 from pebble import ProcessPool
 
-import metisfl.utils.proto_messages_factory as proto_factory
+import metisfl.proto.proto_messages_factory as proto_factory
 from metisfl.learner.grpc_controller_client import GRPCControllerClient
 from metisfl.models.model_dataset import (ModelDataset,
                                           ModelDatasetClassification,
